% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datainput.R
\docType{data}
\name{input}
\alias{input}
\title{Denificja parametrow}
\format{
An object of class \code{data.frame} with 50 rows and 2 columns.
}
\source{
XYZ \url{XYZ}
}
\usage{
input
}
\description{
Denificja parametrow
}
\examples{
\dontrun{
 input


}
}
\keyword{datasets}
