% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dane.R
\docType{data}
\name{wskazniki}
\alias{wskazniki}
\title{Wskazniki emisji transportu drogowego EEA}
\format{

}
\source{

}
\usage{
wskazniki
}
\description{
Wskazniki emisji transportu drogowego EEA
}
\examples{
\dontrun{
 wskazniki
}
}
\keyword{datasets}
