% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WWplot.R
\name{WWplot}
\alias{WWplot}
\title{Funkcja wyswietla wykres na podstawie wyniku funkcji "WWemisja"}
\usage{
WWplot(wynik)
}
\arguments{
\item{wynik}{- data frame, jest to wynik funkcji "WWemisja"}
}
\value{
plot
}
\description{
Przyjmuje nastepujace argumenty:
}
